requirements.txt
# BelgiQ Python dependencies

# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0

# Pipeline / scraping
requests==2.32.3
lxml==5.2.2
pdfplumber==0.11.0
python-dateutil==2.9.0
celery==5.4.0
redis==5.0.7

# AI summarisation (Claude API)
anthropic==0.28.0

# Utilities
rich==13.7.1        # nice terminal output
tenacity==8.3.0     # retry logic for API calls
